feat: add phone field to user registration and profile update; refactor route imports
