+++ backend/src/models/User.models.js
@@ -18,6 +18,7 @@ const userSchema = new Schema(
     name: { type: String, required: true },
     email: { type: String, required: true, unique: true },
     password: { type: String, required: true },
+    phone: { type: String, required: true },
 
     preferences: {
       travelStyle: {
